+++ skate-plugin/project-gen/src/jvmMain/kotlin/slack/tooling/markdown/ui/MarkdownPanel.kt
@@ -131,7 +131,7 @@ private fun jewelMarkdownColor(
   dividerColor: Color = JewelTheme.globalColors.borders.normal,
 ): MarkdownColors {
   val (codeText, codeBackground, inlineCodeText, inlineCodeBackground) =
-    rememberCodeBackground(JewelTheme.globalColors.paneBackground, text)
+    rememberCodeBackground(JewelTheme.globalColors.panelBackground, text)
   return DefaultMarkdownColors(
     text = text,
     codeText = codeText,