+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog
 **Unreleased**
 --------------
 
+0.19.5
+------
+
+_2024-09-23_
+
+- Add workaround for KGP not applying `sourceInformation` compose options in android projects and default it to true.
+
 0.19.4
 ------